me <chrono>

+++ examples/peer_logger.cpp
@@ -87,7 +87,7 @@ static void runEchoServer(Port port) {
         buf[r] = '\0';
         std::cout << "[server] echoing " << r << " byte(s): \"" << buf
                   << "\"\n";
-        conn->send(buf, static_cast<size_t>(static_cast<unsigned int>(r)));
+        conn->send(buf, static_cast<size_t>(r));
     }
 }
 
@@ -141,9 +141,7 @@ static void runUdpPeerLog(Port port) {
     int r = server.receiveFrom(buf, sizeof(buf), from);
     if (r > 0) {
         std::cout << "[udp-server] datagram from " << from.toString() << ": \""
-                  << std::string(
-                         buf, static_cast<size_t>(static_cast<unsigned int>(r)))
-                  << "\"\n";
+                  << std::string(buf, static_cast<size_t>(r)) << "\"\n";
     }
 }
